+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add delay time metric. Exposed to vm as a steal time.
+
+### Changed
+- Rename `wait` metric to `wait_seconds_total`.
 
 ## [2.0.0] - 2021-03-26
 ### Added